refactor oauth token form
